pyproject.toml
[tool.poetry]
name = "fans"
version = "0.1.0"
description = ""
authors = ["Nguyen Van Linh <linh.nguyen04@hastech.vn>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.79.1"}
gunicorn = "^20.1.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.126.0"}
beanie = "^1.11.8"
APScheduler = "^3.9.1"
beautifulsoup4 = "^4.11.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "fans.main:start"
deploy = "fans.main:deploy"